Expose members of `TSHighlighter`.
We at GitHub need this as part of our syntax highlighting pipeline, so as to provide fluent support for highlighting via CSS syntax directives rather than pure-HTML ones. We’re currently hacking around this by keeping a copy of the `TSHighlighter` struct in our source code and doing unsafe casts to it, which is pretty fragile. It would be great to expose these fields (or provide accessor functions).